+++ inst/python/rpytools/call.py
@@ -1,22 +1,23 @@
 import rpycall
 
 
-
-
 def make_python_function(f, name=None):
     def python_function(*args, **kwargs):
-
         # call the function
         value, error = rpycall.call_r_function(f, *args, **kwargs)
 
         if error:
             if isinstance(error, str) and error == "KeyboardInterrupt":
                 # Only reachable if a C++ exception was caught in call_r_function()
-                # otherwise error is always a bool()
+                # otherwise error is always an Exception object
                 raise KeyboardInterrupt()
 
-            raise RuntimeError(value)
+            if isinstance(error, BaseException):
+                raise error
 
+            # basically unreachable since R errors get automatically
+            # converted to python Exceptions, but just in case
+            raise RuntimeError(error)
         return value
 
     if not name is None: