+++ src/reticulate_types.h
@@ -103,6 +103,22 @@ class GILScope {
   }
 };
 
+// define a PythonException struct that we can use to throw an
+// exception from C++ code. The class contains an SEXP with an R condition
+// object that can be used to generate an R error condition.
+struct PythonException {
+  SEXP condition;
+  PythonException(SEXP condition_) : condition(condition_) {}
+};
+
+
+
+// This custom BEGIN_RCPP is effectively identical to upstream
+// except for the last line, which we use to ensure that the
+// GIL is acquired when calling into Python (and released otherwise).
+// Limitations of the macro preprocessor make it difficult to
+// do this in a more elegant way.
+
 #undef BEGIN_RCPP
 #define BEGIN_RCPP                           \
   int rcpp_output_type = 0;                  \
@@ -113,5 +129,55 @@ class GILScope {
   static SEXP stop_sym = Rf_install("stop"); \
   try {                                      \
     GILScope gilscope;
+
+// This custom END_RCPP is effectively identical to upstream
+// except for the addition of one additional catch block, which
+// we use to generate custom R conditions from Python exceptions.
+// Limitations of the macro preprocessor make it difficult to
+// do this in a more elegant way.
+
+#undef END_RCPP
+#define END_RCPP                                                            \
+  }                                                                         \
+  catch (PythonException & __ex__) {                                        \
+    rcpp_output_type = 2;                                                   \
+    rcpp_output_condition = __ex__.condition;                               \
+  }                                                                         \
+  catch (Rcpp::internal::InterruptedException & __ex__) {                   \
+    rcpp_output_type = 1;                                                   \
+  }                                                                         \
+  catch (Rcpp::LongjumpException & __ex__) {                                \
+    rcpp_output_type = 3;                                                   \
+    rcpp_output_condition = __ex__.token;                                   \
+  }                                                                         \
+  catch (Rcpp::exception & __ex__) {                                        \
+    rcpp_output_type = 2;                                                   \
+    rcpp_output_condition = PROTECT(rcpp_exception_to_r_condition(__ex__)); \
+    ++nprot;                                                                \
+  }                                                                         \
+  catch (std::exception & __ex__) {                                         \
+    rcpp_output_type = 2;                                                   \
+    rcpp_output_condition = PROTECT(exception_to_r_condition(__ex__));      \
+    ++nprot;                                                                \
+  }                                                                         \
+  catch (...) {                                                             \
+    rcpp_output_type = 2;                                                   \
+    rcpp_output_condition =                                                 \
+        PROTECT(string_to_try_error("c++ exception (unknown reason)"));     \
+    ++nprot;                                                                \
+  }                                                                         \
+  if (rcpp_output_type == 1) {                                              \
+    Rf_onintr();                                                            \
+  }                                                                         \
+  if (rcpp_output_type == 2) {                                              \
+    SEXP expr = PROTECT(Rf_lang2(stop_sym, rcpp_output_condition));         \
+    ++nprot;                                                                \
+    Rf_eval(expr, R_BaseEnv);                                               \
+  }                                                                         \
+  if (rcpp_output_type == 3) {                                              \
+    Rcpp::internal::resumeJump(rcpp_output_condition);                      \
+  }                                                                         \
+  UNPROTECT(nprot);                                                         \
+  return R_NilValue;
 
 #endif // RETICULATE_TYPES_H